style(suite): refactor assetCard and assetRow so the Coin and Name are separated and reusable in different layouts
